+++ orthanc_api_client/__init__.py
@@ -9,6 +9,7 @@ from .instances_set import InstancesSet
 from .job import Job, JobInfo, JobType, JobStatus
 from .http_client import HttpClient
 from .downloaded_instance import DownloadedInstance
+from .labels_constraint import LabelsConstraint
 
 # __all__ = [
 #     'OrthancApiClient'