+++ orthanc_api_client/labels_constraint.py
@@ -1,0 +1,7 @@
+from strenum import StrEnum
+
+class LabelsConstraint(StrEnum):
+
+    ANY = 'Any'
+    ALL = 'All'
+    NONE = 'None'